⚡️ General Improvements
Added some additional instructional context for those who are very new to using this sort of thing!

Updated the default attributes for <a>.

Links on index.html now include:

`target="_blank"` to automatically open links in a new tab

`rel="noopener"` to navigate to the target resource without granting the new browsing context access to the document that opened it.

The <img> tags on index.html now include a `alt="Text Here"` attribute to help improve accessibility.

LittleLink now scores a 99 in performance, a 94 in accessibility (some brand colors are not accessible, and neither is the LittleLink brand color with light theme), a 100 in best practices, and a 100 in SEO through Google Lighthouse.